+++ R/ggauto_.R
@@ -7,9 +7,13 @@ ggauto_density <- function(var1, var2) {
     p_data <- data.frame(x = var2, y = var1)
     if (is.character(p_data$y)) {
       p_data <- p_data |>
-        dplyr::mutate(y = stringr::str_wrap(stats::reorder(.data$y, .data$x,
-          FUN = stats::median
-        ), 20))
+        dplyr::mutate(
+          y = stats::reorder(.data$y, .data$x, FUN = stats::median),
+          y = factor(
+            stringr::str_wrap(as.character(.data$y), 20),
+            levels = stringr::str_wrap(levels(.data$y), 20)
+          )
+        )
     }
   }
   ggplot2::ggplot(
@@ -39,8 +43,13 @@ ggauto_bar <- function(var1, var2) {
   }
   if (is.character(var1)) {
     p_data <- p_data |>
-      dplyr::mutate(x = stringr::str_wrap(
-        stats::reorder(.data$x, -.data$Count), 20))
+      dplyr::mutate(
+        x = stats::reorder(.data$x, -.data$Count),
+        x = factor(
+          stringr::str_wrap(as.character(.data$x), 20),
+          levels = stringr::str_wrap(levels(.data$x), 20)
+        )
+      )
   }
   g <- ggplot2::ggplot(
     data = p_data,
@@ -96,9 +105,10 @@ ggauto_scatter_facet <- function(var1, var2, var3, base_size) {
   ) +
     auto_zero_line(var2) +
     ggplot2::geom_point(size = 0.2 * base_size, show.legend = FALSE) +
-    ggplot2::scale_colour_manual(values = rep("black", n_cat)) +
+    ggplot2::scale_colour_manual(values = rep(ggplot2::alpha("black", alpha = 0.8), n_cat)) +
+    ggplot2::scale_x_continuous(guide = ggplot2::guide_axis(check.overlap = TRUE)) +
     gghighlight::gghighlight(use_direct_label = FALSE) +
-    ggplot2::facet_wrap(~.data$z) +
+    ggplot2::facet_wrap(~ .data$z) +
     ggplot2::coord_cartesian(expand = FALSE, clip = "off") +
     auto_y_axis(var2)
   return(g)
@@ -122,7 +132,8 @@ ggauto_line_colour <- function(var1, var2, var3, base_size, base_family) {
   p_data <- data.frame(x = var1, y = var2, z = var3)
   last_date <- p_data |>
     dplyr::group_by(.data$z) |>
-    dplyr::slice_max(.data$x)
+    dplyr::slice_max(.data$x) |>
+    dplyr::ungroup()
   g <- ggplot2::ggplot(
     data = p_data,
     mapping = ggplot2::aes(
@@ -147,14 +158,13 @@ ggauto_line_colour <- function(var1, var2, var3, base_size, base_family) {
     ) +
     ggrepel::geom_text_repel(
       data = last_date,
-      mapping = ggplot2::aes(label = .data$z),
+      mapping = ggplot2::aes(label = paste0("  ", .data$z)),
       show.legend = FALSE,
       fontface = "bold",
       family = base_family,
       direction = "y",
       hjust = 0,
       box.padding = 0.5,
-      xlim = c(max(last_date$x), NA),
       seed = 123,
       size = 0.8 * base_size * 0.3528
     ) +
@@ -166,7 +176,16 @@ ggauto_line_colour <- function(var1, var2, var3, base_size, base_family) {
 #' @noRd
 ggauto_line_facet <- function(var1, var2, var3, base_size) {
   p_data <- data.frame(x = var1, y = var2, z = var3)
-  n_cat <- length(unique(var3))
+  if (is.factor(var3)) {
+    facet_levels <- levels(var3)
+  } else {
+    non_na <- p_data[!is.na(p_data$y), ]
+    non_na <- non_na[order(non_na$x), ]
+    last_vals <- non_na[!duplicated(non_na$z, fromLast = TRUE), ]
+    facet_levels <- last_vals$z[order(last_vals$y)]
+  }
+  p_data$z <- factor(p_data$z, levels = facet_levels)
+  n_cat <- length(facet_levels)
   g <- ggplot2::ggplot(
     data = p_data,
     mapping = ggplot2::aes(
@@ -177,10 +196,14 @@ ggauto_line_facet <- function(var1, var2, var3, base_size) {
     auto_zero_line(var2) +
     ggplot2::geom_line(
       show.legend = FALSE,
-      linewidth = 0.1 * base_size
+      linewidth = 0.08 * base_size
     ) +
     ggplot2::scale_colour_manual(values = rep("black", n_cat)) +
-    ggplot2::scale_y_continuous(labels = scales::comma) +
+    ggplot2::scale_x_date(guide = ggplot2::guide_axis(check.overlap = TRUE)) +
+    ggplot2::scale_y_continuous(
+      labels = scales::comma,
+      guide = ggplot2::guide_axis(check.overlap = TRUE)
+    ) +
     gghighlight::gghighlight(use_direct_label = FALSE) +
     ggplot2::facet_wrap(~ .data$z) +
     ggplot2::coord_cartesian(expand = FALSE, clip = "off")
@@ -199,16 +222,23 @@ ggauto_heatmap <- function(var1, var2, var3, base_size) {
   }
   if (is.character(var1)) {
     p_data <- p_data |>
-      dplyr::mutate(x = stringr::str_wrap(
-        stats::reorder(.data$x, .data$n,
-        FUN = sum
-      ), 20))
+      dplyr::mutate(
+        x = stats::reorder(.data$x, .data$n, FUN = sum),
+        x = factor(
+          stringr::str_wrap(as.character(.data$x), 20),
+          levels = stringr::str_wrap(levels(.data$x), 20)
+        )
+      )
   }
   if (is.character(var2)) {
     p_data <- p_data |>
-      dplyr::mutate(y = stringr::str_wrap(stats::reorder(.data$y, -.data$n,
-        FUN = sum
-      ), 20))
+      dplyr::mutate(
+        y = stats::reorder(.data$y, -.data$n, FUN = sum),
+        y = factor(
+          stringr::str_wrap(as.character(.data$y), 20),
+          levels = stringr::str_wrap(levels(.data$y), 20)
+        )
+      )
   }
   label_in_data <- p_data |>
     dplyr::filter(.data$r > 0.24)