+++ NEWS.md
@@ -1,5 +1,15 @@
 # ggauto
 
+## ggauto 0.0.2
+
+* Add CRAN badge
+* Fix: character ordering
+* Fix: allow specification of only title or subtitle
+* Improve facet labelling and axis text
+* Improve facet ordering and appearance
+* Fix: `coord_flip()` issues with axis label
+* Remove `penguins` examples
+
 ## ggauto 0.0.1
 
 * Initialise package